Add content hash as query string to static file paths
