+++ src/Elastic.Markdown/Slices/Index.cshtml
@@ -16,7 +16,8 @@
 		StaticUrlPathPrefix = Model.StaticUrlPathPrefix,
 		GithubEditUrl = Model.GithubEditUrl,
 		AllowIndexing = Model.AllowIndexing,
-		Features = Model.Features
+		Features = Model.Features,
+		GetContentHashFunc = Model.GetContentHashFunc
 	};
 }
 <section id="elastic-docs-v3">
